no more UnkStarFuncs.inc.c
